Multi-angle car sprites
- Image asset added for all car rotation angles in img/res and AssetDefines.h. Also removed old car sprite sheet.
- Updated Car struct to use new rotation angles in Car.cpp and Assets.cpp.
- Fixed a minor bug with Car::cy() not reading viewportRect height (it read Ypos instead, ooooops).
- Angle for car selection now set based on car's anglesprite value in OnLoop.cpp.
- Car anglesprite value now set during PlayerCarsUpdate() in OnLoop.cpp.
- Resolved some conversion warnings in OnLoop.cpp and BoostStreak.cpp.
- Disabled debug lines on active car and ball in OnRender.cpp.

+++ Missile Champions/Objects/BoostStreak.cpp
@@ -3,11 +3,11 @@
 void BoostStreak::SpawnSprite(double _x, double _y, double _angle, Assets* assets) {
 	this->x = _x;
 	this->y = _y;
-	for (double a = 11.25, i = 0; a <= 371.25; a += 22.5, i++) {
+	for (double a = 11.25, i = 0.0; a <= 371.25; a += 22.5, i++) {
 		if (_angle < a && _angle >= (a - 22.5)) {
-			angleSprite = i;
+			angleSprite = (int)i;
 		}
-		if (i == 7) i = -1;
+		if (i == 7.0) i = -1.0;
 	}
 	this->timeAlive = 250;
 	this->decaySprite = 0;
